+++ run-tests.sh
@@ -17,7 +17,6 @@ sbcl --load src/core/version.lisp \
      --load src/backend/6502.lisp \
      --load src/backend/6510.lisp \
      --load src/backend/45gs02.lisp \
-     --load src/emit/output.lisp \
      --load src/backend/65c02.lisp \
      --load src/backend/r65c02.lisp \
      --load src/backend/65816.lisp \
@@ -33,6 +32,7 @@ sbcl --load src/core/version.lisp \
      --load src/debugger/6502.lisp \
      --load src/optimizer/6502.lisp \
      --load src/optimizer/65c02.lisp \
+     --load src/emit/output.lisp \
      --load tests/test-debugger-6502.lisp \
      --load tests/test-65c02.lisp \
      --load tests/test-r65c02.lisp \
@@ -61,6 +61,7 @@ sbcl --load src/core/version.lisp \
      --load tests/test-linker-6502.lisp \
      --load tests/test-optimizer.lisp \
      --load tests/test-restarts.lisp \
+     --load tests/test-listing.lisp \
      --load tests/run-tests.lisp \
      --eval "(cl-asm/test:run-all-tests)" \
      --quit